Make AudioParameters a class instead of a struct
Also collapses some long parameter lists into AudioParameters and moves some
of the hardcoded values (e.g. 16 bit audio in AudioDevice) to more appropriate
locations.

BUG=115902
TEST=manually testing everything works out

Review URL: https://chromiumcodereview.appspot.com/9655018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128054 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/audio/mac/audio_low_latency_input_mac.h
@@ -66,7 +66,7 @@ class AUAudioInputStream : public AudioInputStream {
   virtual double GetVolume() OVERRIDE;
 
   // Returns the current hardware sample rate for the default input device.
-  MEDIA_EXPORT static double HardwareSampleRate();
+  MEDIA_EXPORT static int HardwareSampleRate();
 
   bool started() const { return started_; }
   AudioUnit audio_unit() { return audio_unit_; }
